# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Bonsai Journal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /trees:
    get:
      operationId: listTrees
      tags: [trees]
      summary: List all trees with optional filters
      parameters:
        - name: search
          in: query
          schema: { type: string }
        - name: climate
          in: query
          schema: { type: string }
        - name: foliage
          in: query
          schema: { type: string }
        - name: stage
          in: query
          schema: { type: string }
        - name: status
          in: query
          schema: { type: string }
        - name: tag
          in: query
          schema: { type: string }
        - name: tags
          in: query
          description: "Comma-separated list of tags (OR logic). Passed as string[], serialized by URL builder."
          schema:
            type: array
            items: { type: string }
          style: form
          explode: false
        - name: limit
          in: query
          schema: { type: integer }
        - name: offset
          in: query
          schema: { type: integer }
      responses:
        "200":
          description: List of trees
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tree"
    post:
      operationId: createTree
      tags: [trees]
      summary: Add a new tree
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TreeInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tree"
  /trees/{id}:
    get:
      operationId: getTree
      tags: [trees]
      summary: Get a single tree
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Tree detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tree"
        "404":
          description: Not found
    patch:
      operationId: updateTree
      tags: [trees]
      summary: Update a tree
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TreeUpdate"
      responses:
        "200":
          description: Updated tree
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tree"
        "404":
          description: Not found
    delete:
      operationId: deleteTree
      tags: [trees]
      summary: Delete a tree
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "204":
          description: Deleted
  /trees/{id}/logs:
    get:
      operationId: listTreeLogs
      tags: [logs]
      summary: Get care logs for a tree
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: List of care logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CareLog"
    post:
      operationId: createTreeLog
      tags: [logs]
      summary: Add a care log entry
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CareLogInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CareLog"
  /trees/{id}/logs/{logId}:
    patch:
      operationId: updateTreeLog
      tags: [logs]
      summary: Update a care log entry
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
        - name: logId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CareLogInput"
      responses:
        "200":
          description: Updated log
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CareLog"
    delete:
      operationId: deleteTreeLog
      tags: [logs]
      summary: Delete a care log entry
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
        - name: logId
          in: path
          required: true
          schema: { type: string }
      responses:
        "204":
          description: Deleted
  /trees/{id}/reminders:
    get:
      operationId: listTreeReminders
      tags: [reminders]
      summary: Get reminders for a tree
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: List of reminders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CareReminder"
    post:
      operationId: createTreeReminder
      tags: [reminders]
      summary: Add a reminder for a tree
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CareReminderInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CareReminder"
  /trees/{id}/reminders/{reminderId}:
    patch:
      operationId: updateTreeReminder
      tags: [reminders]
      summary: Update a reminder (e.g. mark complete)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
        - name: reminderId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CareReminderUpdate"
      responses:
        "200":
          description: Updated reminder
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CareReminder"
    delete:
      operationId: deleteTreeReminder
      tags: [reminders]
      summary: Delete a reminder
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
        - name: reminderId
          in: path
          required: true
          schema: { type: string }
      responses:
        "204":
          description: Deleted
  /trees/{id}/timeline:
    get:
      operationId: getTreeTimeline
      tags: [trees]
      summary: Combined timeline of care logs and reminders for a tree
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Timeline events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TimelineEvent"
...
  /trees/{id}/photos:
    get:
      operationId: listTreePhotos
      tags: [trees]
      summary: Get progression photos for a tree
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: List of progression photos ordered by date
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TreePhoto"
    post:
      operationId: createTreePhoto
      tags: [trees]
      summary: Add a progression photo to a tree
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TreePhotoInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TreePhoto"
        "404":
          description: Tree not found
  /trees/{id}/photos/{photoId}:
    patch:
      operationId: updateTreePhoto
      tags: [trees]
      summary: Update a progression photo's date
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
        - name: photoId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TreePhotoUpdate"
      responses:
        "200":
          description: Updated photo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TreePhoto"
        "404":
          description: Not found
    delete:
      operationId: deleteTreePhoto
      tags: [trees]
      summary: Delete a progression photo
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
        - name: photoId
          in: path
          required: true
          schema: { type: string }
      responses:
        "204":
          description: Deleted
components:
  schemas:
...
    Tree:
      type: object
      required: [id, name, tags, createdAt, updatedAt]
      properties:
        id:
          type: string
        name:
          type: string
        species:
          type: ["string", "null"]
        acquiredDate:
          type: ["string", "null"]
        climate:
          type: ["string", "null"]
        foliage:
          type: ["string", "null"]
        style:
          type: ["string", "null"]
        stage:
          type: ["string", "null"]
        status:
          type: ["string", "null"]
        tags:
          type: array
          items: { type: string }
        notes:
          type: ["string", "null"]
        photoUrl:
          type: ["string", "null"]
        coverThumb:
          type: ["string", "null"]
        coverPosition:
          type: ["object", "null"]
          required: [x, y, zoom]
          properties:
            x:
              type: number
            y:
              type: number
            zoom:
              type: number
        createdAt:
          type: string
        updatedAt:
          type: string
    TreeInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        species:
          type: string
        acquiredDate:
          type: string
        climate:
          type: string
        foliage:
          type: string
        style:
          type: string
        stage:
          type: string
        status:
          type: string
        tags:
          type: array
          items: { type: string }
        notes:
          type: string
        photoUrl:
          type: string
        coverThumb:
          type: string
    TreeUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        species:
          type: string
        acquiredDate:
          type: string
        climate:
          type: string
        foliage:
          type: string
        style:
          type: string
        stage:
          type: string
        status:
          type: string
        tags:
          type: array
          items: { type: string }
        notes:
          type: string
        photoUrl:
          type: string
        coverThumb:
          type: string
        coverPosition:
          type: object
          required: [x, y, zoom]
          properties:
            x:
              type: number
            y:
              type: number
            zoom:
              type: number
    CareLog:
      type: object
      required: [id, treeId, type, date, createdAt]
      properties:
        id:
          type: string
        treeId:
          type: string
        type:
          type: string
        date:
          type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    CareLogInput:
      type: object
      required: [type, date]
      properties:
        type:
          type: string
        date:
          type: string
        notes:
          type: string
    CareReminder:
      type: object
      required: [id, treeId, type, dueDate, completed, createdAt]
      properties:
        id:
          type: string
        treeId:
          type: string
        type:
          type: string
        dueDate:
          type: string
        notes:
          type: ["string", "null"]
        completed:
          type: boolean
        createdAt:
          type: string
    CareReminderInput:
      type: object
      required: [type, dueDate]
      properties:
        type:
          type: string
        dueDate:
          type: string
        notes:
          type: string
    CareReminderUpdate:
      type: object
      properties:
        type:
          type: string
        dueDate:
          type: string
        notes:
          type: string
        completed:
          type: boolean
...
    TimelineEvent:
      type: object
      required: [id, kind, date, type]
      properties:
        id:
          type: string
        kind:
          type: string
          description: "log or reminder"
        date:
          type: string
        type:
          type: string
        notes:
          type: ["string", "null"]
        completed:
          type: ["boolean", "null"]
    TreePhoto:
      type: object
      required: [id, treeId, photoUrl, takenAt, createdAt]
      properties:
        id:
          type: string
        treeId:
          type: string
        photoUrl:
          type: string
        photoThumb:
          type: ["string", "null"]
        takenAt:
          type: string
        createdAt:
          type: string
    TreePhotoInput:
      type: object
      required: [photoUrl]
      properties:
        photoUrl:
          type: string
        photoThumb:
          type: string
        takenAt:
          type: string
    TreePhotoUpdate:
      type: object
      required: [takenAt]
      properties:
        takenAt:
          type: string